+++ crates/rune-macros/src/context.rs
@@ -617,19 +617,23 @@ impl Context {
         };
 
         Tokens {
+            any_type_info: quote!(#module::runtime::AnyTypeInfo),
             any: quote!(#module::Any),
+            compile_error: quote!(#module::compile::Error),
             context_error: quote!(#module::compile::ContextError),
             from_value: quote!(#module::runtime::FromValue),
+            full_type_of: quote!(#module::runtime::FullTypeOf),
             hash: quote!(#module::Hash),
             id: quote!(#module::parse::Id),
-            install_with: quote!(#module::compile::InstallWith),
+            install_with: quote!(#module::module::InstallWith),
             macro_context: quote!(#module::macros::MacroContext),
-            module: quote!(#module::compile::Module),
+            maybe_type_of: quote!(#module::runtime::MaybeTypeOf),
+            module_variant: quote!(#module::module::Variant),
+            module: quote!(#module::module::Module),
             named: quote!(#module::compile::Named),
             object: quote!(#module::runtime::Object),
             opaque: quote!(#module::parse::Opaque),
             option_spanned: quote!(#module::ast::OptionSpanned),
-            compile_error: quote!(#module::compile::Error),
             parse: quote!(#module::parse::Parse),
             parser: quote!(#module::parse::Parser),
             pointer_guard: quote!(#module::runtime::SharedPointerGuard),
@@ -637,47 +641,47 @@ impl Context {
             raw_into_mut: quote!(#module::runtime::RawMut),
             raw_into_ref: quote!(#module::runtime::RawRef),
             raw_str: quote!(#module::runtime::RawStr),
+            result: quote!(::core::result::Result),
             shared: quote!(#module::runtime::Shared),
             span: quote!(#module::ast::Span),
             spanned: quote!(#module::ast::Spanned),
             to_tokens: quote!(#module::macros::ToTokens),
             to_value: quote!(#module::runtime::ToValue),
             token_stream: quote!(#module::macros::TokenStream),
+            try_result: quote!(#module::runtime::try_result),
             tuple: quote!(#module::runtime::Tuple),
             type_info: quote!(#module::runtime::TypeInfo),
-            any_type_info: quote!(#module::runtime::AnyTypeInfo),
+            type_name: quote!(::core::any::type_name),
             type_of: quote!(#module::runtime::TypeOf),
-            maybe_type_of: quote!(#module::runtime::MaybeTypeOf),
-            full_type_of: quote!(#module::runtime::FullTypeOf),
             unit_struct: quote!(#module::runtime::UnitStruct),
             unsafe_from_value: quote!(#module::runtime::UnsafeFromValue),
             unsafe_to_value: quote!(#module::runtime::UnsafeToValue),
             value: quote!(#module::runtime::Value),
             variant_data: quote!(#module::runtime::VariantData),
-            compile_variant: quote!(#module::compile::Variant),
             vm_error: quote!(#module::runtime::VmError),
             vm_result: quote!(#module::runtime::VmResult),
-            try_result: quote!(#module::runtime::try_result),
-            type_name: quote!(::core::any::type_name),
-            result: quote!(::core::result::Result),
         }
     }
 }
 
 pub(crate) struct Tokens {
+    pub(crate) any_type_info: TokenStream,
     pub(crate) any: TokenStream,
+    pub(crate) compile_error: TokenStream,
     pub(crate) context_error: TokenStream,
     pub(crate) from_value: TokenStream,
+    pub(crate) full_type_of: TokenStream,
     pub(crate) hash: TokenStream,
     pub(crate) id: TokenStream,
     pub(crate) install_with: TokenStream,
     pub(crate) macro_context: TokenStream,
+    pub(crate) maybe_type_of: TokenStream,
+    pub(crate) module_variant: TokenStream,
     pub(crate) module: TokenStream,
     pub(crate) named: TokenStream,
     pub(crate) object: TokenStream,
     pub(crate) opaque: TokenStream,
     pub(crate) option_spanned: TokenStream,
-    pub(crate) compile_error: TokenStream,
     pub(crate) parse: TokenStream,
     pub(crate) parser: TokenStream,
     pub(crate) pointer_guard: TokenStream,
@@ -685,29 +689,25 @@ pub(crate) struct Tokens {
     pub(crate) raw_into_mut: TokenStream,
     pub(crate) raw_into_ref: TokenStream,
     pub(crate) raw_str: TokenStream,
+    pub(crate) result: TokenStream,
     pub(crate) shared: TokenStream,
     pub(crate) span: TokenStream,
     pub(crate) spanned: TokenStream,
     pub(crate) to_tokens: TokenStream,
     pub(crate) to_value: TokenStream,
     pub(crate) token_stream: TokenStream,
+    pub(crate) try_result: TokenStream,
     pub(crate) tuple: TokenStream,
     pub(crate) type_info: TokenStream,
-    pub(crate) any_type_info: TokenStream,
+    pub(crate) type_name: TokenStream,
     pub(crate) type_of: TokenStream,
-    pub(crate) maybe_type_of: TokenStream,
-    pub(crate) full_type_of: TokenStream,
     pub(crate) unit_struct: TokenStream,
     pub(crate) unsafe_from_value: TokenStream,
     pub(crate) unsafe_to_value: TokenStream,
     pub(crate) value: TokenStream,
     pub(crate) variant_data: TokenStream,
-    pub(crate) compile_variant: TokenStream,
     pub(crate) vm_error: TokenStream,
     pub(crate) vm_result: TokenStream,
-    pub(crate) try_result: TokenStream,
-    pub(crate) type_name: TokenStream,
-    pub(crate) result: TokenStream,
 }
 
 impl Tokens {
