+++ crates/rune/src/compile.rs
@@ -59,21 +59,6 @@ pub mod meta;
 pub(crate) use self::meta::{Doc, ItemMeta};
 pub use self::meta::{MetaRef, SourceMeta};
 
-mod function_meta;
-pub(crate) use self::function_meta::{AssociatedFunctionData, FunctionData, IterFunctionArgs};
-pub use self::function_meta::{
-    AssociatedFunctionKind, AssociatedFunctionName, FunctionMeta, MacroMeta, ToFieldFunction,
-    ToInstance,
-};
-#[doc(hidden)]
-pub use self::function_meta::{FunctionMetaData, FunctionMetaKind, MacroMetaData, MacroMetaKind};
-
-pub(crate) mod module;
-pub use self::module::{
-    AssocType, AssociatedFunctionKey, AsyncFunction, AsyncInstFn, Function, InstFn, InstallWith,
-    Module, ModuleFunction, Variant,
-};
-
 mod pool;
 pub(crate) use self::pool::{ItemId, ModId, ModMeta, Pool};
