+++ crates/rune/src/runtime/bytes.rs
@@ -10,7 +10,8 @@ use crate::no_std::prelude::*;
 
 use serde::{Deserialize, Serialize};
 
-use crate::compile::{InstallWith, Named};
+use crate::compile::Named;
+use crate::module::InstallWith;
 use crate::runtime::{
     FromValue, Mut, RawMut, RawRef, RawStr, Ref, UnsafeFromValue, Value, VmResult,
 };
